restructure; add edit support
restructured so get_multi_input is its own file and changed start_menu to only have menu functions (hence the rename). get_multi_input was changed such that it can now edit strings using left and right (as well as tackling adding and deleting characters whilst in the middle of the string). get_multi_input was also cleaned up to be more clear